geo_prior/web_app/requirements.txt
Flask==1.0.2
numpy==1.16.4
torch==1.2.0
Pillow==6.1.0
Jinja2==2.10.1
